requirements.txt
absl-py==2.3.1
accelerate==1.13.0
acme==1.21.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosmtplib==5.1.0
alabaster==0.7.12
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.12.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
Authlib==1.6.9
Automat==20.2.0
av==17.0.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
boto3==1.42.28
botocore==1.42.28
brotli==1.2.0
bs4==0.0.2
cached_path==1.8.1
cachetools==6.2.6
certbot==1.21.0
certbot-dns-cloudflare==1.18.0
certifi==2026.2.25
cffi==2.0.0
cftime==1.6.5
chardet==4.0.0
charset-normalizer==3.4.4
click==8.3.2
cloud-init==25.1.4
cloudflare==2.8.14
cmudict==1.1.3
cn2an==0.5.22
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.2
cryptography==46.0.5
cssselect2==0.9.0
ctranslate2==4.7.1
cuda-bindings==12.9.4
cuda-pathfinder==1.4.3
curl_cffi==0.13.0
cycler==0.12.1
datasets==4.8.3
dateparser==1.1.8
dbus-python==1.2.18
decorator==5.2.1
Deprecated==1.3.1
dill==0.4.1
Distance==0.1.3
distro==1.7.0
distro-info===1.1build1
dnspython==2.8.0
docopt==0.6.2
docutils==0.17.1
easyocr==1.7.2
ecdsa==0.19.1
email-validator==2.3.0
emoji==1.2.0
eng_to_ipa==0.0.2
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fail2ban==0.11.2
fastapi==0.135.1
fastapi-mail==1.6.2
faster-whisper==1.2.1
feedparser==6.0.12
ffmpy==1.0.0
filelock==3.25.2
Flask==3.1.2
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.2.0
fugashi==1.3.0
fuzzywuzzy==0.18.0
g2p-en==2.1.0
g2pkk==0.1.2
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
gitsome==0.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-genai==1.65.0
google-generativeai==0.8.6
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gradio==6.3.0
gradio_client==2.0.3
greenlet==3.3.2
groovy==0.1.2
grpcio==1.76.0
grpcio-status==1.71.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
hangul-jamo==1.0.1
hf-xet==1.4.2
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.11
ImageIO==2.37.3
imagesize==1.3.0
importlib_metadata==9.0.0
importlib_resources==6.5.2
incremental==21.3.0
inflect==7.0.0
iniconfig==2.3.0
itsdangerous==2.2.0
jaconv==0.4.1
jamo==0.4.1
jeepney==0.7.1
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
josepy==1.10.0
jpype1==1.6.0
jsonlines==1.2.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keras==3.12.1
keyring==23.5.0
kiwisolver==1.4.9
kobart @ git+https://github.com/SKT-AI/KoBART@79ff1d76ed47550695a05f4cd923642d754b0929
kollocate==0.0.2
konlpy==0.6.0
koparadigm==0.10.0
kss==6.0.6
langid==1.1.6
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-loader==0.5
Levenshtein==0.27.3
libclang==18.1.1
libhwp==0.2.0
librosa==0.9.1
llvmlite==0.46.0
logging==0.4.9.6
loguru==0.7.2
lxml==4.8.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mecab-python3==1.0.9
-e git+https://github.com/myshell-ai/MeloTTS.git@209145371cff8fc3bd60d7be902ea69cbdb7965a#egg=melotts
ml_dtypes==0.5.4
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.7.1
multipledispatch==1.0.0
multiprocess==0.70.19
multitasking==0.0.12
namex==0.1.0
narwhals==2.17.0
netCDF4==1.7.4
netifaces==0.11.0
networkx==2.8.8
ninja==1.13.0
nltk==3.9.1
num2words==0.5.12
numba==0.63.1
numpy==1.26.4
numpydoc==1.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
olefile==0.47
onnxruntime==1.23.2
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.19.0
orjson==3.11.5
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
parsedatetime==2.6
passlib==1.7.4
pdfminer.six==20251230
pdfplumber==0.11.9
pecab==1.0.8
peewee==4.0.1
pexpect==4.8.0
pillow==12.1.0
plac==1.4.5
platformdirs==4.5.1
plotly==6.6.0
pluggy==1.6.0
ply==3.11
pooch==1.8.2
proces==0.1.7
prompt-toolkit==3.0.28
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1-modules==0.2.1
pyclipper==1.4.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
pydyf==0.12.1
Pygments==2.19.2
PyGObject==3.42.1
PyHamcrest==2.0.2
pyhwp==0.1b15
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pykakasi==2.2.1
pyOpenSSL==26.0.0
pyparsing==3.3.2
pypdfium2==5.6.0
pyphen==0.17.2
pypinyin==0.50.0
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
pytest==9.0.2
python-apt==2.4.0+ubuntu4.1
python-bidi==0.6.7
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
python-debian===0.1.43ubuntu1
python-docx==0.8.11
python-dotenv==1.2.1
python-jose==3.5.0
python-Levenshtein==0.27.3
python-magic==0.4.24
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2
python-multipart==0.0.21
pytz==2026.1.post1
PyYAML==6.0.3
RapidFuzz==3.14.3
redis==7.3.0
regex==2026.4.4
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==0.9.1
resampy==0.4.3
rich==13.9.4
roman==3.3
rsa==4.9.1
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
SecretStorage==3.3.1
semantic-version==2.10.0
sentence-transformers==5.3.0
sentencepiece==0.2.1
service-identity==18.1.0
setproctitle==1.2.2
sgmllib3k==1.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
solapi==5.0.3
sortedcontainers==2.4.0
sos==4.4
soundfile==0.13.1
soupsieve==2.8.1
soxr==1.0.0
Sphinx==4.3.2
SQLAlchemy==2.0.48
srt==3.5.3
ssh-import-id==5.11
starlette==0.50.0
streamlit==1.54.0
stripe==14.4.1
sympy==1.14.0
systemd-python==234
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.5.1
tinyhtml5==2.1.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.4.0
tomlkit==0.13.3
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tornado==6.5.4
tossi==0.3.1
tqdm==4.67.3
transformers==4.27.4
trio==0.33.0
trio-websocket==0.12.2
triton==3.6.0
twilio==9.10.4
Twisted==22.1.0
txtsplit==1.0.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==202603
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.7
unidic==1.1.0
unidic-lite==1.0.8
uritemplate==4.2.0
urllib3==1.26.20
user-agents==2.2.0
uvicorn==0.41.0
uvloop==0.22.1
vosk==0.3.45
wadllib==1.3.6
wasabi==0.10.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.5
weasyprint==68.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.5
Whoosh==2.7.4
wrapt==2.0.1
wsproto==1.3.2
xkit==0.0.0
xlrd==1.2.0
xonsh==0.11.0
xxhash==3.6.0
yarl==1.23.0
yfinance==1.2.0
zipp==3.23.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
zopfli==0.4.1
